Build multiple Golang versions only for amd64 version (#252) (#250)

+++ helm-charts/scripts/install-toolchain.sh
@@ -23,30 +23,42 @@ ARCH=$([[ $(uname -m) = "x86_64" ]] && echo 'amd64' || echo 'arm64')
 TMP_DIR="${BUILD_DIR}/tmp"
 mkdir -p "${TOOLS_DIR}"
 
-HELM_VERSION="v3.4.1"
-KUBECTL_VERSION=v1.19.6
-KIND_VERSION=v0.5.1
-RELEASE_BRANCH=1-19
-RELEASE=1
+HELM_VERSION=v3.7.1
+KUBECTL_VERSION=v1.20.7
+KIND_VERSION=v0.11.1
+RELEASE_BRANCH=1-20
+RELEASE=8
 
 ## Install kubectl
-curl -sSL "https://distro.eks.amazonaws.com/kubernetes-${RELEASE_BRANCH}/releases/${RELEASE}/artifacts/kubernetes/${KUBECTL_VERSION}/bin/linux/amd64/kubectl" -o "${TOOLS_DIR}/kubectl"
-chmod +x "${TOOLS_DIR}/kubectl"
+if ! command -v kubectl &> /dev/null; then
+    echo "kubectl could not be found. Downloading..."
+    curl -sSL "https://distro.eks.amazonaws.com/kubernetes-${RELEASE_BRANCH}/releases/${RELEASE}/artifacts/kubernetes/${KUBECTL_VERSION}/bin/linux/amd64/kubectl" -o "${TOOLS_DIR}/kubectl"
+    chmod +x "${TOOLS_DIR}/kubectl"
+fi
 
 ## Install kubeval
-mkdir -p "${TMP_DIR}/kubeval"
-curl -sSL https://github.com/instrumenta/kubeval/releases/latest/download/kubeval-${PLATFORM}-${ARCH}.tar.gz | tar xz -C "${TMP_DIR}/kubeval"
-mv "${TMP_DIR}/kubeval/kubeval" "${TOOLS_DIR}/kubeval"
+if ! command -v kubeval &> /dev/null; then
+    echo "kubeval could not be found. Downloading..."
+    mkdir -p "${TMP_DIR}/kubeval"
+    curl -sSL https://github.com/instrumenta/kubeval/releases/latest/download/kubeval-${PLATFORM}-${ARCH}.tar.gz | tar xz -C "${TMP_DIR}/kubeval"
+    mv "${TMP_DIR}/kubeval/kubeval" "${TOOLS_DIR}/kubeval"
+fi
 
 ## Install helm
-mkdir -p "${TMP_DIR}/helm"
-curl -sSL https://get.helm.sh/helm-${HELM_VERSION}-${PLATFORM}-${ARCH}.tar.gz | tar xz -C "${TMP_DIR}/helm"
-mv "${TMP_DIR}/helm/${PLATFORM}-${ARCH}/helm" "${TOOLS_DIR}/helm"
-rm -rf "${PLATFORM}-${ARCH}"
+if ! command -v helm &> /dev/null; then
+    echo "helm could not be found. Downloading..."
+    mkdir -p "${TMP_DIR}/helm"
+    curl -sSL https://get.helm.sh/helm-${HELM_VERSION}-${PLATFORM}-${ARCH}.tar.gz | tar xz -C "${TMP_DIR}/helm"
+    mv "${TMP_DIR}/helm/${PLATFORM}-${ARCH}/helm" "${TOOLS_DIR}/helm"
+    rm -rf "${PLATFORM}-${ARCH}"
+fi
 
 ## Install kind
-curl -sSL "https://github.com/kubernetes-sigs/kind/releases/download/${KIND_VERSION}/kind-${PLATFORM}-${ARCH}" -o "${TOOLS_DIR}/kind"
-chmod +x "${TOOLS_DIR}/kind"
+if ! command -v kind &> /dev/null; then
+    echo "kind could not be found. Downloading..."
+    curl -sSL "https://github.com/kubernetes-sigs/kind/releases/download/${KIND_VERSION}/kind-${PLATFORM}-${ARCH}" -o "${TOOLS_DIR}/kind"
+    chmod +x "${TOOLS_DIR}/kind"
+fi
 
 rm -rf ${TMP_DIR}
